modpack/pyproject.toml
[project]
name = "modpack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.66.0",
    "loguru>=0.7.3",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer[all]>=0.17.3",
    "uv>=0.8.15",
]
